Add camera fragment to scan barcodes and refactoring

+++ app/src/main/java/com/philipgurr/smartshoppinglist/util/extensions/ShoppingListId.kt
@@ -1,5 +1,5 @@
 package com.philipgurr.smartshoppinglist.util.extensions
 
 fun String.toId(): String {
-    return hashCode().toString()
+    return hashCode().toUInt().toString()
 }
